Added: rotatePattern function to check all possible rotations of a pattern flipHorizontal function to check horizontal flips of a pattern flipVertical function to check vertical flips of a pattern isGliderOrLWSS function to check for gliders and LWSS checkForStableSpaceship to ensure spaceships meet stable generation requirements.
Updated:
matchesPattern function to support varients of patterns and threading to improve performance.
isBlinkerOrToad removed extra phase of blinker as it is handled by rotation
isBlockOrBeehive removed vertical beehive varient as it is handled by rotation
runExperiment updated to support searching for gliders or LWSS
updated to break early if a pattern is found
updated to allow users to determine the max amount of experiments and break when that limit is reached to prevent infinite loop is pattern is not found.